// server/api/errors.ts
export enum UserErrors {
  NO_USER_FOUND = "no-user-found",
  WRONG_CREDENTIALS = "wrong-credentials",
  USERNAME_ALREADY_EXISTS = "username-already-exists",
  EMAIL_ALREADY_EXISTS = "email-already-exists",
  CURRENT_PASSWORD_INCORRECT = "Current password is incorrect'"
}


export enum ProductErrors {
NO_PRODUCT_FOUND = "no-product=found",
NOT_ENOUGH_STOCK = "not-enough=stock",
NO_AVAILABLE_MONEY = "no-available-money",
}